moved NullLiteral
