+++ src_smartcontract/CMakeLists.txt
@@ -13,6 +13,7 @@ add_subdirectory(sc_declare_types)
 add_subdirectory(sc_expression)
 add_subdirectory(sc_expression_arithmetic)
 add_subdirectory(sc_expression_bit)
+add_subdirectory(sc_expression_literal)
 add_subdirectory(sc_expression_logical)
 add_subdirectory(sc_statement)
 add_subdirectory(sc_statement_ctrl)
