+++ src_smartcontract/sc/CodeElement.cpp
@@ -59,11 +59,13 @@
 #include "sc_expression/LiteralExpression.h"
 #include "sc_expression/NumberLiteral.h"
 #include "sc_expression/BooleanLiteral.h"
-#include "sc_expression/NullLiteral.h"
 #include "sc_expression/ParenthesisExpression.h"
 #include "sc_expression/VariableIdentifier.h"
-
 #include "sc_expression/FunctionCallExpression.h"
+
+#include "sc_expression_literal/NullLiteral.h"
+
+
 #include "sc_expression_bit/AndExpression.h"
 #include "sc_expression_bit/BitReverseExpression.h"
 #include "sc_expression_bit/OrExpression.h"
