+++ src_smartcontract/sc/alinouslang.h
@@ -61,10 +61,11 @@
 #include "sc_expression/FunctionCallExpression.h"
 #include "sc_expression/LiteralExpression.h"
 #include "sc_expression/MemberReferenceExpression.h"
-#include "sc_expression/NullLiteral.h"
 #include "sc_expression/NumberLiteral.h"
 #include "sc_expression/ParenthesisExpression.h"
 #include "sc_expression/VariableIdentifier.h"
+
+#include "sc_expression_literal/NullLiteral.h"
 
 #include "sc_expression_arithmetic/AddExpression.h"
 #include "sc_expression_arithmetic/MultiplicativeExpression.h"
