+++ src_smartcontract/sc_expression_literal/CMakeLists.txt
@@ -1,0 +1,8 @@
+
+
+set(__src
+	NullLiteral.cpp
+)
+
+handle_sub(codablecashlib "${__src}" smartcontract sc_expression_literal)
+
