+++ src_smartcontract/sc_expression_literal/NullLiteral.h
